+++ packages/react/src/experimental/OneDataCollection/__stories__/full-height.stories.tsx
@@ -20,14 +20,41 @@ type Story = StoryObj<typeof meta>
 export const Basic: Story = {
   parameters: {
     chromatic: { disableSnapshot: true },
-    layout: "fullscreen",
   },
+  decorators: [
+    (Story) => (
+      <div style={{ height: "400px", width: "100%", display: "flex" }}>
+        <Story />
+      </div>
+    ),
+  ],
   render: () => <ExampleComponent frozenColumns={2} fullHeight />,
 }
 
 export const WithPagination: Story = {
+  ...Basic,
+  render: () => {
+    const paginatedMockUsers = generateMockUsers(50)
+    const dataAdapter = createDataAdapter({
+      data: paginatedMockUsers,
+      delay: 500,
+      paginationType: "pages",
+    })
+
+    return (
+      <ExampleComponent
+        frozenColumns={2}
+        fullHeight
+        dataAdapter={dataAdapter}
+      />
+    )
+  },
+}
+
+export const WithPaginationAndGrouping: Story = {
+  ...WithPagination,
   parameters: {
-    layout: "fullscreen",
+    chromatic: { disableSnapshot: false },
   },
   render: () => {
     const paginatedMockUsers = generateMockUsers(50)
@@ -41,6 +68,17 @@ export const WithPagination: Story = {
       <ExampleComponent
         frozenColumns={2}
         fullHeight
+        selectable={(item) => item.index}
+        grouping={{
+          collapsible: true,
+          mandatory: true,
+          groupBy: {
+            department: {
+              name: "Department",
+              label: (groupId) => groupId,
+            },
+          },
+        }}
         dataAdapter={dataAdapter}
       />
     )